vim: enable "clipboard clipboard" for gvim
My real need is that vim messes with no clipboard by default, and copies
to the "clipboard clipboard" (not the primary clipboard) when I
explicitly request it. I could not quikly find out how to do that, so is
a workaround instead: do this depending on if I'm in vim or gvim.